First implementation of radial filling, missing proper yaml integration and seems unoptimized when using glHandler->dispatchShader (70s instead of 40s)

+++ DataStructures.cpp
@@ -108,7 +108,7 @@ heightMap * emptyHeightMapFromPointGrid(pointGrid *grid) {
     return new heightMap{.heights = std::vector<float>(grid->resolutionX * grid->resolutionY),
             .resolutionX = grid->resolutionX,
             .resolutionY = grid->resolutionY,
-            .dataSize = static_cast<long long>(sizeof(float) * grid->resolutionX * grid->resolutionY),
+            .dataSize = static_cast<long>(sizeof(float) * grid->resolutionX * grid->resolutionY),
             .min = grid->min,
             .max = grid->max};
 }
